Make VS Code use Git installed under WSL
By default, VS Code makes use of Git installed under Windows which
causes changes to the file permissions even though WSL mounts the linux
filesystem with its own file permissions. andy-5/wslgit forwards all
arguments to git running in WSL:

https://github.com/andy-5/wslgit

Just a few days ago, Microsoft announced that they are working on Remote
Development using VS Code which, together with WSL 2.0 which is
scheduled to be released in June of this year dramatically improves
working on source code that lives under WSL or inside a container.
Awesome stuff!

https://github.com/Microsoft/vscode/issues/9502#issuecomment-488792093

As with any new technology, it is not without bugs so it is worth
waiting a little while longer before adopting as the comment below
points out:

https://github.com/Microsoft/vscode/issues/9502#issuecomment-490032893