Configure vscode, eslint, prettier and vetur
All of these tools offer something I want but they also offer overlapping
formatting and linting functionality to conflict with one another. VSCode
leaves the linting and formatting to ESLint now. Vetur is merely included for
syntax highlighting of .vue files. It could serve as a nice one stop solution
for Vue development in the future but for now, using it is too much of a pain.